// games-BE-ac-v2/libs/shared/enums/gameBeResultCodes.enum.ts
export enum GameBeResultCodes {
  INTERNAL_SERVER_ERROR = 2000,

  RGS_ERROR_INVALID_TOKEN = 2001,
  RGS_ERROR_INIT_REQUEST_FAILED = 2002,
  RGS_ERROR_BET_REQUEST_FAILED = 2003,
  RGS_ERROR_REFUND_REQUEST_FAILED = 2004,
  RGS_ERROR_WIN_REQUEST_FAILED = 2005,

  TOKEN_ISSUED = 2006,
  NO_TOKEN = 2007,
  INVALID_TOKEN = 2008,
  NO_PLAYER = 2009,

  INVALID_GUESS_HILO = 2010,
  INVALID_CASHOUT_HILO = 2011,
  BET_OPEN_HILO = 2012,
  INVALID_BET_HILO = 2013,
  INVALID_SKIP_HILO = 2014,

  BET_OPEN_MINES = 2015,
  INVALID_BET_MINES = 2016,
  FIELD_SELECTED_MINES = 2017,
  INVALID_CASHOUT_MINES = 2018,
  INVALID_GAMECODE_MINES = 2019,
}